requirements.txt
langchain
pypdf2
python-dotenv
streamlit
langchain-openai
langchain-text-splitters
langchain-community
faiss-cpu